backend/requirements.txt
fastapi
uvicorn[standard]
nba_api
python-dotenv
